solution/pyproject.toml
[project]
name = "Advertising_platform"
version = "1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "colorlog>=6.9.0",
 "django-environ>=0.11.2",
 "django-extensions>=3.2.3",
 "django-ninja>=1.3.0",
 "gunicorn>=23.0.0",
 "psycopg2-binary>=2.9.10",
 "python-json-logger>=3.2.1",
 "dj_database_url",
 "ruff>=0.9.6",
 "django-redis>=5.4.0",
 "django-prometheus>=2.3.1",
 "minio>=7.2.15",
 "django-minio-storage>=0.5.7",
 "pillow>=11.1.0",
 "yandex-cloud-ml-sdk>=0.3.1",
]

[tool.ruff]
builtins = []
cache-dir = ".ruff_cache"
exclude = [
 ".bzr",
 ".direnv",
 ".eggs",
 ".git",
 ".git-rewrite",
 ".hg",
 ".mypy_cache",
 ".nox",
 ".pants.d",
 ".pytype",
 ".ruff_cache",
 ".svn",
 ".tox",
 ".venv",
 "__pypackages__",
 "_build",
 "buck-out",
 "dist",
 "migrations",
 "node_modules",
 "venv",
]
extend-exclude = []
extend-include = []
fix = false
fix-only = false
force-exclude = true
include = ["*.py", "*.pyi", "*.ipynb", "**/pyproject.toml"]
indent-width = 4
line-length = 79

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 79
exclude = []
indent-style = "space"
line-ending = "lf"
preview = false
quote-style  = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
exclude = ["tests.py"]
fixable = ["ALL"]
ignore = ["ARG", "COM812", "D", "ISC001", "PT009" ,"N813", "RUF012", "ANN201", "RUF001"]
select = ["ALL"]
task-tags = ["TODO"]
